Split converters.py into helpers.py, registry.py, and converters.py
- helpers.py: shared utilities, matmul helpers, var/param dict builders
- registry.py: atom converter functions and ATOM_CONVERTERS dict
- converters.py: convert_expr entry point and param-aware matmul/multiply
